requirements-pinned.txt
# Dependency pinning - Freeze exact versions để reproducible builds
# Sử dụng: pip install -r requirements-pinned.txt

# === FRAMEWORK ===
fastapi==0.109.0
uvicorn[standard]==0.27.0
starlette==0.35.0

# === STREAMLIT ===
streamlit==1.30.0

# === DATABASE ===
mysql-connector-python==8.3.0
SQLAlchemy==2.0.23  # optional - for ORM

# === FACE RECOGNITION ===
deepface==0.0.75
facenet-pytorch==2.5.0
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
Pillow==10.2.0

# === DEEP LEARNING ===
torch==2.1.2  # CPU version; use torch==2.1.2+cu118 for GPU
torchvision==0.16.2
scikit-learn==1.3.2
scikit-image==0.22.0

# === DATA PROCESSING ===
numpy==1.26.3
pandas==2.2.0

# === HTTP CLIENT ===
httpx==0.25.2
requests==2.31.0

# === UTILITIES ===
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# === LOGGING ===
python-json-logger==2.0.7

# === CACHING ===
redis==5.0.1  # optional

# === TESTING ===
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # for testing

# === CODE QUALITY ===
black==23.12.1
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1
